# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse — Agentic Maturity Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engagements/{id}/members:
    get:
      operationId: listEngagementMembers
      tags: [engagements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EngagementMember" }
    post:
      operationId: addEngagementMember
      tags: [engagements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddEngagementMemberInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EngagementMember" }
  /engagements/{id}/members/{memberId}:
    delete:
      operationId: removeEngagementMember
      tags: [engagements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: memberId, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Removed
  /engagements:
    get:
      operationId: listEngagements
      tags: [engagements]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Engagement" }
    post:
      operationId: createEngagement
      tags: [engagements]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEngagementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Engagement" }
  /engagements/{id}:
    get:
      operationId: getEngagement
      tags: [engagements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Engagement" }
    patch:
      operationId: updateEngagement
      tags: [engagements]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEngagementInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Engagement" }
  /engagements/{id}/dashboard:
    get:
      operationId: getEngagementDashboard
      tags: [dashboard]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EngagementDashboard" }
  /engagements/{id}/activity:
    get:
      operationId: getEngagementActivity
      tags: [dashboard]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - name: kind
          in: query
          required: false
          description: Comma-separated list of activity kinds to include.
          schema: { type: string }
        - name: severity
          in: query
          required: false
          schema: { type: string, enum: [info, critical] }
        - name: actor
          in: query
          required: false
          description: Case-insensitive substring match against actor name or email.
          schema: { type: string }
        - name: from
          in: query
          required: false
          description: ISO timestamp lower bound (inclusive).
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          description: ISO timestamp upper bound (inclusive).
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityEvent" }
  /engagements/{id}/activity.csv:
    get:
      operationId: exportEngagementActivityCsv
      tags: [dashboard]
      summary: Admin-only CSV export of the audit log (same filters as GET /activity).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: kind, in: query, required: false, schema: { type: string } }
        - { name: severity, in: query, required: false, schema: { type: string, enum: [info, critical] } }
        - { name: actor, in: query, required: false, schema: { type: string } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
        "403":
          description: Admin role required
...
components:
  schemas:
...
    EngagementStatus:
      type: string
      enum: [draft, active, collecting, ready_for_analysis, exported, archived]
...
    DeliverableStatus:
      type: string
      enum: [draft, reviewed, locked]
...
    EngagementMember:
      type: object
      required: [id, engagementId, email, role, status, createdAt]
      properties:
        id: { type: string }
        engagementId: { type: string }
        userId: { type: string, nullable: true }
        email: { type: string }
        name: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        role: { type: string, enum: [owner, assessor, viewer] }
        status: { type: string, enum: [pending, active] }
        invitedBy: { type: string, nullable: true }
        createdAt: { type: string }
    AddEngagementMemberInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { type: string, enum: [owner, assessor, viewer] }
    Engagement:
      type: object
      required:
        [id, clientName, sponsor, teamCount, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        clientName: { type: string }
        sponsor: { type: string }
        teamCount: { type: integer }
        scope: { type: string }
        teams:
          type: array
          items: { type: string }
        kickoffDate: { type: string, nullable: true }
        targetDeliveryDate: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/EngagementStatus" }
        modules:
          type: array
          items: { type: string, enum: [security, data, platform, design] }
        createdAt: { type: string }
        updatedAt: { type: string }
    CreateEngagementInput:
      type: object
      required: [clientName, sponsor, teamCount]
      properties:
        clientName: { type: string }
        sponsor: { type: string }
        teamCount: { type: integer, minimum: 1 }
        scope: { type: string }
        teams:
          type: array
          items: { type: string }
        kickoffDate: { type: string, nullable: true }
        targetDeliveryDate: { type: string, nullable: true }
        modules:
          type: array
          items: { type: string, enum: [security, data, platform, design] }
    UpdateEngagementInput:
      type: object
      properties:
        clientName: { type: string }
        sponsor: { type: string }
        teamCount: { type: integer }
        scope: { type: string }
        teams:
          type: array
          items: { type: string }
        kickoffDate: { type: string, nullable: true }
        targetDeliveryDate: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/EngagementStatus" }
        modules:
          type: array
          items: { type: string, enum: [security, data, platform, design] }
    EngagementDashboard:
      type: object
      required:
        [
          engagementId,
          surveyResponseRate,
          surveySent,
          surveyCompleted,
          interviewsCompleted,
          interviewsTotal,
          artifactCount,
          connectorsHealthy,
          connectorsTotal,
          daysToTarget,
          deliverableStatuses,
        ]
      properties:
        engagementId: { type: string }
        surveyResponseRate: { type: number }
        surveySent: { type: integer }
        surveyCompleted: { type: integer }
        interviewsCompleted: { type: integer }
        interviewsTotal: { type: integer }
        artifactCount: { type: integer }
        connectorsHealthy: { type: integer }
        connectorsTotal: { type: integer }
        daysToTarget: { type: integer, nullable: true }
        evidenceCount: { type: integer }
        deliverableStatuses:
          type: object
          required: [heatmap, gapAnalysis, actionPlan, entryPoint, npv]
          properties:
            heatmap: { $ref: "#/components/schemas/DeliverableStatus" }
            gapAnalysis: { $ref: "#/components/schemas/DeliverableStatus" }
            actionPlan: { $ref: "#/components/schemas/DeliverableStatus" }
            entryPoint: { $ref: "#/components/schemas/DeliverableStatus" }
            npv: { $ref: "#/components/schemas/DeliverableStatus" }
    ActivityEvent:
      type: object
      required: [id, kind, message, createdAt, severity]
      properties:
        id: { type: string }
        kind: { type: string }
        message: { type: string }
        createdAt: { type: string }
        actorUserId: { type: string, nullable: true }
        actorName: { type: string, nullable: true }
        actorEmail: { type: string, nullable: true }
        actorAvatarUrl: { type: string, nullable: true }
        severity:
          type: string
          enum: [info, critical]
        payload:
          type: object
          additionalProperties: true
          description: Structured before/after detail captured by the emitting route.
        requestId:
          type: string
          nullable: true
          description: Correlation id matching the x-request-id response header and pino logs.
...