# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse — Agentic Maturity Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventPayload } from "./activityEventPayload";
import type { ActivityEventSeverity } from "./activityEventSeverity";

export interface ActivityEvent {
  id: string;
  kind: string;
  message: string;
  createdAt: string;
  actorUserId?: string | null;
  actorName?: string | null;
  actorEmail?: string | null;
  actorAvatarUrl?: string | null;
  severity: ActivityEventSeverity;
  /** Structured before/after detail captured by the emitting route. */
  payload?: ActivityEventPayload;
  /** Correlation id matching the x-request-id response header and pino logs. */
  requestId?: string | null;
}